feat: Implement unified build for correct CSS output
This commit consolidates the build process and refines Tailwind CSS
configuration to ensure `dist/index.css` correctly includes all
necessary styles for the library, while relying on the host
application for Tailwind utilities and theming.

Key Changes:

1.  **Unified Build Process:**
    *   Reverted to a single `vite build` command (using the main
        `vite.config.ts`) for building both JavaScript and CSS.
    *   Removed the separate `vite.config.css.js` and associated
        `build:css` script from `package.json`.
    *   `src/index.ts` now re-imports `src/assets/library.css` to ensure
        these custom global styles are part of the main CSS bundle.

2.  **Mode-Dependent Tailwind Configuration in `vite.config.ts`:**
    *   For development (`yarn dev`): Tailwind scans all project files
        and Preflight is enabled, ensuring `src/assets/style.css`
        provides a full styling environment. This should fix issues
        with styles not appearing in local development.
    *   For library build (`yarn build`):
        *   `content` array is configured to scan Vue component paths
            (e.g., `./src/components/**/*.vue`). This allows Tailwind
            to process `@import "tailwindcss"` and `@apply` directives
            within component `<style scoped>` blocks (like in `DXButton.vue`),
            ensuring their structural styles are compiled and included
            in `dist/index.css`.
        *   `corePlugins: { preflight: false }` is set to prevent
            global Preflight rules from Tailwind being included in the
            library's `dist/index.css`.

3.  **Expected CSS Output (`dist/index.css`):**
    *   Should now contain both the plain, `.vue-dynamix`-scoped custom
        styles from `src/assets/library.css` (which includes `plugin.css`
        and `card.css`).
    *   Should also include the compiled, `data-v-xxxxxx`-scoped CSS
        from Vue components' `<style scoped>` blocks (e.g., the `.btn-*`
        styles from `DXButton.vue`).
    *   Should *not* include global Preflight or a general dump of all
        Tailwind utilities not specifically used by internal `@apply`s.

4.  **Plain CSS:**
    *   `@apply` directives in `src/assets/plugin.css` and
        `src/assets/card.css` were previously converted to plain CSS.
        These files are imported by `library.css`.

This set of changes aims to ensure that `dist/index.css` contains all
necessary styles for `vue-dynamix` to function correctly when used as
a package (including component-specific structural styles like `.btn-*`),
while the host application provides the Tailwind theme and utility class
definitions for classes used in component templates.
