+++ src/components/DXChat/DXChatLayout/DXChatLayout.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import {nextTick, onMounted, ref} from "vue";
-import {vScrollbar} from "@/directives/scrollbar.ts";
 import type {ChatLayoutProps} from "@/types/chat";
 
 
@@ -73,12 +72,13 @@ defineExpose({
 		<slot name="subheader"/>
 	</div>
 
-	<div
-		ref="bodyElementRef"
-		:class="[bodyClass, 'vue-dynamix']"
-		v-scrollbar="scrollable ? scrollSize : undefined"
-		@scroll="onScroll"
-	>
+	<div ref="bodyElementRef" :class="[
+				bodyClass,
+				scrollable ? 'scroll' : '',
+				scrollSize ? `scroll-${scrollSize}` : '',
+				'vue-dynamix'
+			]"
+	     @scroll="onScroll">
 		<slot/>
 	</div>
 	<div ref="footerElementRef" :class="[footerClass, 'vue-dynamix']">
