+++ src/directives/directives.ts
@@ -1,9 +1,7 @@
 import {vRipple} from "./ripple";
 import {vTimeUpdate} from "./timeUpdate";
-import {vScrollbar} from "./scrollbar";
 
 export  {
   vRipple,
-  vTimeUpdate,
-  vScrollbar
+  vTimeUpdate
 }
