+++ src/index.ts
@@ -1,3 +1,5 @@
+import './assets/library.css';
+
 // Composables
 export { useParsedOptions } from "@/composables/useParsedOptions";
 export { useForm } from "@/composables/useForm";
